Introduce a new variable to the legend
In this case we have positive or negative series, and a new state for each kind called validated.

So to distinguish well that state we created four series, two positives with the validated and not validated values and the same for the negative ones.

The problem here is with the legend. JQPlot always takes the colors from the seriesColors properties, so to be well we need to show there the four colors and the corresponding colors for the negatives series.